refactor: extracts BackgroundLayer to its own file
