+++ Sources/Clappr_iOS/Classes/Base/Layers/LayersCompositor.swift
@@ -1,16 +1,4 @@
-import Foundation
-
-final class BackgroundLayer: UIView {
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        backgroundColor = UIColor.black
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
+import UIKit
 
 class LayersCompositor {
